style(bgp): the lint findings the tree carried are gone
`assert.Equal(t, uint8(message.NotifyCeaseBFDDown), ...)` converts a
value that is already uint8, twice. unconvert has said so for as long
as the two lines have existed, and they were the LAST two lint findings
in the tree: `./le verify lint run` now reports 0 issues on all
eighteen flavors, which it has not done at any point in this session.
Both lines sit inside RFC-tagged tests, so only the owner approves the
edit, and the two RFC-approved trailers carry his ruling. No assertion
moves: the line above each one reads `uint8(message.NotifyCease)`,
which IS a conversion because NotifyCease is a NotifyErrorCode, and the
asymmetry between a typed error code and a bare uint8 subcode is what
made the redundant one look deliberate.
The structural override names the red this commit REMOVES. The cached
failure index was written at 04:56 and gives the lint stage one related
path, which is this file.
Measured before reaching for it: typing the Cease subcodes so the
conversion becomes necessary is 73 uses across 27 files, 15 of them
tests in the most heavily tagged area of the BGP tree. That route needs
more approvals than it avoids.

RFC-approved: reactor.TestNotificationRefusedBySocketStillRecordsTheReason: Thomas ruled on 2026-09-21, answering a list that named this edit and what it changes: his answer was to proceed with all six. The edit deletes a uint8 conversion around message.NotifyCeaseBFDDown, which is already uint8; no assertion moves.
RFC-approved: reactor.TestNotificationDeliveredIsNotRecordedAsUnsent: Thomas ruled on 2026-09-21, answering a list that named this edit and what it changes: his answer was to proceed with all six. The edit deletes a uint8 conversion around message.NotifyCeaseBFDDown, which is already uint8; no assertion moves.

+++ internal/component/bgp/reactor/peer_last_error_test.go
@@ -74,7 +74,7 @@ func TestNotificationRefusedBySocketStillRecordsTheReason(t *testing.T) {
 	stats := peer.Stats()
 	assert.Equal(t, plugin.NotifSendFailed, stats.LastNotifDirection)
 	assert.Equal(t, uint8(message.NotifyCease), stats.LastNotifCode)
-	assert.Equal(t, uint8(message.NotifyCeaseBFDDown), stats.LastNotifSubcode)
+	assert.Equal(t, message.NotifyCeaseBFDDown, stats.LastNotifSubcode)
 	assert.False(t, stats.LastNotifTime.IsZero())
 
 	// Nothing reached the peer, so nothing counts as sent, and the flag that
@@ -114,7 +114,7 @@ func TestNotificationDeliveredIsNotRecordedAsUnsent(t *testing.T) {
 	stats = peer.Stats()
 	assert.Equal(t, plugin.NotifSent, stats.LastNotifDirection)
 	assert.Equal(t, uint8(message.NotifyCease), stats.LastNotifCode)
-	assert.Equal(t, uint8(message.NotifyCeaseBFDDown), stats.LastNotifSubcode)
+	assert.Equal(t, message.NotifyCeaseBFDDown, stats.LastNotifSubcode)
 	assert.Equal(t, uint32(1), stats.NotificationsSent)
 	assert.True(t, peer.notificationExchanged.Load())
 }
